% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySummary.R
\name{mySummary}
\alias{mySummary}
\title{Provides a detailed summary of a numeric vector or variable}
\usage{
mySummary(var, na.rm = TRUE, ...)
}
\arguments{
\item{var}{a numeric variable}

\item{na.rm}{if TRUE (default), it removes the missing values}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data frame of summary measures
}
\description{
Provides a detailed summary of a numeric vector or variable
}
\examples{

Dat <- data.frame(age = c(40, 35, 27, 52, 29, 29, 46),
sex = c("M", "F", "M", "F", "F", "M", "M"))

mySummary(Dat$age)

mySummary(Dat[Dat$sex=="M", ]$age)






}
